requirements.txt
Django
djangorestframework
Pillow
psycopg2-binary
python-decouple
drf_yasg